Restructue native renderer entry
Renamed ReactNative.js to ReactNativeStack.js
ReactNative.js is the new entry which checks if it is gonna use stack or
fiber. Check is done using ReactNativeFeatureFlags.useFiber
